+++ vidsjs/modules/index.js
@@ -8,3 +8,4 @@ exports.view = require('./view');
 exports.seen = require('./seen');
 exports.users = require('./users');
 exports.loginchecker = require('./loginchecker');
+exports.deleted = require('./deleted');